+++ anjuyi1/Decorate/HouseInspect/FreeOfferViewController.m
@@ -442,7 +442,7 @@
         
     }
     else{
-        [codeBtn setTitle:[NSString stringWithFormat:@"%ld S",time] forState:(UIControlStateNormal)];
+        [codeBtn setTitle:[NSString stringWithFormat:@"%ld S",(long)time] forState:(UIControlStateNormal)];
     }
 }
