+++ anjuyi1/Decorate/Master/MasterListViewController.m
@@ -98,7 +98,7 @@
                           @"level":_level,
                           @"min":_min,
                           @"max":_max,
-                          @"page":[NSString stringWithFormat:@"%ld",self.page]};
+                          @"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
     
     [MBProgressHUD showHUDAddedTo:self.view animated:YES];
     
@@ -150,7 +150,7 @@
                            @"level":_level,
                            @"min":_min,
                            @"max":_max,
-                           @"page":[NSString stringWithFormat:@"%ld",self.page]};
+                           @"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
 
     
     
@@ -255,7 +255,7 @@
 
 - (void)selectIndex:(NSInteger)index{
     
-    NSLog(@"%ld",index);
+    NSLog(@"%ld",(long)index);
     
     if (index == 3) {
