+++ anjuyi1/Home/Charging/ChargingHomeViewController.m
@@ -73,7 +73,7 @@
 - (void)pullDownRefresh{
     NSString *path = [NSString stringWithFormat:@"%@/charging/article_list",KURL];
     
-    NSDictionary *dic = @{@"page":[NSString stringWithFormat:@"%ld",self.page]};
+    NSDictionary *dic = @{@"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
     
     [MBProgressHUD showHUDAddedTo:self.view animated:YES];
     
@@ -113,7 +113,7 @@
 - (void)pullUpLoadMore{
     NSString *path = [NSString stringWithFormat:@"%@/charging/article_list",KURL];
     
-    NSDictionary *dic = @{@"page":[NSString stringWithFormat:@"%ld",self.page]};
+    NSDictionary *dic = @{@"page":[NSString stringWithFormat:@"%ld",(long)self.page]};
     
     [MBProgressHUD showHUDAddedTo:self.view animated:YES];
     
@@ -236,7 +236,7 @@
         [back setTag:i];
         [self.headerView addSubview:back];
         
-        [back addSubview:[Tools creatImage:CGRectMake((KScreenWidth/4 - 25)/2, 23, 25, 25) image:[NSString stringWithFormat:@"cdz_ico%ld",i+1]]];
+        [back addSubview:[Tools creatImage:CGRectMake((KScreenWidth/4 - 25)/2, 23, 25, 25) image:[NSString stringWithFormat:@"cdz_ico%ld",(long)(i+1)]]];
         
         [back addSubview:[Tools creatLabel:CGRectMake(0, 50, KScreenWidth/4, 12) font:[UIFont systemFontOfSize:12] color:[UIColor colorWithHexString:@"#3b3b3b"] alignment:(NSTextAlignmentCenter) title:tArr[i]]];
